Refactor project strucutre.
