+++ src/components/messages_components/MessageListText.vue
